refactor: merge taxes from delivery note to Sales Invoice
